pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-summarizer-agent"
version = "1.0.0"
description = "Turn any webpage into structured, summarized information that other agents can use instantly"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Rediet Dagnew", email = "redietdagnew@example.com"}
]
keywords = ["ai", "agent", "summarization", "web-scraping", "nlp", "meshcore"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "readability-lxml>=0.8.1",
    "anthropic>=0.18.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
    "types-beautifulsoup4>=4.12.0",
]

[project.urls]
Homepage = "https://github.com/redietdagnew/web-summarizer-agent"
Documentation = "https://github.com/redietdagnew/web-summarizer-agent#readme"
Repository = "https://github.com/redietdagnew/web-summarizer-agent"
Issues = "https://github.com/redietdagnew/web-summarizer-agent/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=web_summarizer --cov-report=term-missing --cov-report=html"
